backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.26
pytest>=7.0
pytest-django>=4.5
mypy>=1.2
ruff>=0.0.281
black>=23.0
isort>=5.0
jpholiday
